Restructure for better readability, separate scene functionality in operators, improve label creation to the point where it's actually accurate no matter what
