+++ src/dataset_generation/operations/__init__.py
@@ -1,5 +1,4 @@
 from .scene_clearer import SceneClearer
-from .obj_importer import ObjImporter
-from .surface_map_generator import SurfaceMapGenerator
-from .crack_generator import CrackGenerator
 from .material_loader import MaterialLoader
+from .crack_renderer import CrackRenderer
+from .label_thresholder import LabelThresholder
