+++ src/dataset_generation/operations/material_loader.py
@@ -8,20 +8,23 @@ class MaterialLoader:
 
     KEYWORD_BRICK = 'brick'
     KEYWORD_MORTAR = 'mortar'
-    KEYWORD_CRACK = 'crack'
+    KEYWORD_FOREGROUND = 'foreground'
 
     BRICK_MATERIALS = [
         'Brick - red',
         'Brick - gray'
     ]
     MORTAR_MATERIALS = ['Mortar - white']
-    CRACK_MATERIALS = ['Crack marker']
+    FOREGROUND_MATERIAL = 'Marker - foreground'
 
     def __call__(self) -> dict[str, dict[str, bpy.types.Material]]:
+        """
+        Load all materials.
+        """
         materials = {
             self.KEYWORD_BRICK: {},
             self.KEYWORD_MORTAR: {},
-            self.KEYWORD_CRACK: {}
+            self.KEYWORD_FOREGROUND: bpy.data.materials[self.FOREGROUND_MATERIAL]
         }
 
         for material_name in self.BRICK_MATERIALS:
@@ -30,8 +33,5 @@ class MaterialLoader:
         for material_name in self.MORTAR_MATERIALS:
             if material_name in bpy.data.materials:
                 materials[self.KEYWORD_MORTAR][material_name] = bpy.data.materials[material_name]
-        for material_name in self.CRACK_MATERIALS:
-            if material_name in bpy.data.materials:
-                materials[self.KEYWORD_CRACK][material_name] = bpy.data.materials[material_name]
 
         return materials
