+++ src/dataset_generation/operations/scene_clearer.py
@@ -1,19 +1,19 @@
 import bpy
 
+SAFE_COLLECTION: str = 'Base assets'
+
 
 class SceneClearer:
     """
     Operation that clears the scene aside from a single marked collection
     """
 
-    SAFE_COLLECTION: str = 'Base assets'
-
     def __call__(self) -> None:
         """
         Clear objects from the scene aside from the safe collection
         """
         for collection in bpy.data.collections:
-            if collection.name != self.SAFE_COLLECTION:
+            if collection.name != SAFE_COLLECTION:
                 for obj in collection.objects:
                     bpy.data.objects.remove(obj, do_unlink=True)
                 bpy.data.collections.remove(collection, do_unlink=True)
@@ -21,7 +21,7 @@ class SceneClearer:
         # Remove lingering objects
         for obj in bpy.context.scene.collection.objects:
             bpy.data.objects.remove(obj, do_unlink=True)
-        
+
         # Remove any unused garbage
         for iterable in [bpy.data.meshes, bpy.data.cameras, bpy.data.objects]:
             for iter_object in iterable:
