+++ src/requirements.txt
@@ -7,5 +7,3 @@ opencv-python==4.8.1.78
 scikit-learn==1.3.1
 scikit-image==0.22.0
 fake-bpy-module-3.4
-Pillow==10.0.1
-tk-tools
